requirements.txt
arxiv
PyYAML
pydantic
chromadb
tqdm
numpy
psutil
pypdf
langgraph
langchain-ollama
langchain